requirements.txt
fastapi
uvicorn
python-multipart

torch
torchaudio
transformers

openai-whisper
pyannote.audio

librosa
soundfile
ffmpeg-python
numpy
scipy

tqdm
pandas